chore: feedDetail protocol 파일 생성
