+++ KnockKnock-iOS/Scenes/Feed/FeedDetail/FeedDetailWorker.swift
@@ -7,55 +7,6 @@
 
 import Foundation
 
-protocol FeedDetailWorkerProtocol {
-  func getFeedDetail(feedId: Int, completionHandler: @escaping (ApiResponse<FeedDetail>?) -> Void)
-
-  func checkTokenIsValidated() async -> Bool
-  
-  func requestLike(
-    isLike: Bool,
-    feedId: Int,
-    completionHandler: @escaping (ApiResponse<Bool>?) -> Void
-  )
-
-  func toggleLike(feedDetail: FeedDetail?) -> FeedDetail?
-  func fetchLikeList(
-    feedId: Int,
-    completionHandler: @escaping (ApiResponse<[Like.Info]>?) -> Void
-  )
-
-  func getAllComments(
-    feedId: Int,
-    completionHandler: @escaping (ApiResponse<[Comment]>?) -> Void
-  )
-  func fetchVisibleComments(comments: [Comment]?) -> [Comment]
-
-  func requestAddComment(
-    comment: AddCommentDTO,
-    completionHandler: @escaping (ApiResponse<Bool>?) -> Void
-  )
-  func requestDeleteComment(
-    feedId: Int,
-    commentId: Int,
-    comments: [Comment],
-    completionHandler: @escaping (ApiResponse<Bool>?) -> Void
-  )
-
-  func requestDeleteFeed(
-    feedId: Int,
-    completionHandler: @escaping (ApiResponse<Bool>?) -> Void
-  )
-  func requestHidePost(
-    feedId: Int,
-    completionHandler: @escaping (ApiResponse<Bool>?) -> Void
-  )
-  func requestReportFeed(
-    feedId: Int,
-    reportType: ReportType,
-    completionHandler: @escaping (ApiResponse<Bool>?) -> Void
-  )
-}
-
 final class FeedDetailWorker: FeedDetailWorkerProtocol {
   typealias OnCompletionHandler = (ApiResponse<Bool>?) -> Void
